+++ persona26/persona26.php
@@ -1,11 +1,12 @@
 <?php
 /**
  * Plugin Name: Persona26
- * Plugin URI: https://github.com/cchatterton/persona26
+ * Plugin URI: https://github.com/cchatterton/persona26/releases/latest
  * Description: Extends IA with cookie and tracks related sessions
- * Version: 0.2
+ * Version: 0.3
  * Requires at least: 6.0
  * Requires PHP: 8.1
+ * Update URI: https://github.com/cchatterton/persona26
  * Author: Techn
  * Author URI: https://techn.com.au
  * Text Domain: persona26
@@ -15,7 +16,7 @@ if (!defined('ABSPATH')) {
     exit;
 }
 
-define('P26_VERSION', '0.2');
+define('P26_VERSION', '0.3');
 define('P26_PLUGIN_FILE', __FILE__);
 define('P26_PLUGIN_DIR', plugin_dir_path(__FILE__));
 define('P26_PLUGIN_URL', plugin_dir_url(__FILE__));